requirements.txt
Flask==1.1.1
numpy==1.17.0
Pillow==6.1.0
opencv-python
